feat: use same top heading for insight card and exported insight
