+++ frontend/src/lib/components/Cards/InsightCard/InsightCard.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx'
 import { BindLogic, useActions, useValues } from 'kea'
-import { capitalizeFirstLetter, dateFilterToText } from 'lib/utils'
+import { capitalizeFirstLetter } from 'lib/utils'
 import React, { useEffect, useState } from 'react'
 import { Layout } from 'react-grid-layout'
 import {
@@ -34,7 +34,6 @@ import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
 import { ResizeHandle1D, ResizeHandle2D } from '../handles'
 import './InsightCard.scss'
 import { InsightDetails } from './InsightDetails'
-import { INSIGHT_TYPES_METADATA, InsightTypeMetadata, QUERY_TYPES_METADATA } from 'scenes/saved-insights/SavedInsights'
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { ActionsHorizontalBar, ActionsLineGraph, ActionsPie } from 'scenes/trends/viz'
 import { DashboardInsightsTable } from 'scenes/insights/views/InsightsTable/DashboardInsightsTable'
@@ -62,11 +61,11 @@ import {
 import { CardMeta, Resizeable } from 'lib/components/Cards/CardMeta'
 import { DashboardPrivilegeLevel, FEATURE_FLAGS } from 'lib/constants'
 import { Query } from '~/queries/Query/Query'
-import { containsHogQLQuery, dateRangeFor, isDataTableNode, isInsightQueryNode } from '~/queries/utils'
 import { PieChartFilled } from '@ant-design/icons'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { QueriesUnsupportedHere } from 'lib/components/Cards/InsightCard/QueriesUnsupportedHere'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { TopHeading } from 'lib/components/Cards/InsightCard/TopHeading'
 
 type DisplayedType = ChartDisplayType | 'RetentionContainer' | 'FunnelContainer' | 'PathsContainer'
 
@@ -414,44 +413,6 @@ function InsightMeta({
                 </>
             }
         />
-    )
-}
-
-function TopHeading({ insight }: { insight: InsightModel }): JSX.Element {
-    const { filters, query } = insight
-
-    let insightType: InsightTypeMetadata
-
-    // check the query first because the backend still adds defaults to empty filters :/
-    if (!!query?.kind) {
-        if (isDataTableNode(query) && containsHogQLQuery(query)) {
-            insightType = QUERY_TYPES_METADATA[query.source.kind]
-        } else {
-            insightType = QUERY_TYPES_METADATA[query.kind]
-        }
-    } else if (!!filters.insight) {
-        insightType = INSIGHT_TYPES_METADATA[filters.insight]
-    } else {
-        // maintain the existing default
-        insightType = INSIGHT_TYPES_METADATA[InsightType.TRENDS]
-    }
-
-    let { date_from, date_to } = filters
-    if (!!query) {
-        const queryDateRange = dateRangeFor(query)
-        if (!!queryDateRange) {
-            date_from = queryDateRange.date_from
-            date_to = queryDateRange.date_to
-        }
-    }
-
-    const defaultDateRange = query == undefined || isInsightQueryNode(query) ? 'Last 7 days' : null
-    const dateText = dateFilterToText(date_from, date_to, defaultDateRange)
-    return (
-        <>
-            <span title={insightType?.description}>{insightType?.name}</span>
-            {dateText ? <> • {dateText}</> : null}
-        </>
     )
 }
